+++ extensions/cornerstone/src/panels/PanelSegmentation.tsx
@@ -2,12 +2,10 @@ import React from 'react';
 import { SegmentationTable } from '@ohif/ui-next';
 import { useActiveViewportSegmentationRepresentations } from '../hooks/useActiveViewportSegmentationRepresentations';
 import { metaData } from '@cornerstonejs/core';
+import { useSystem } from '@ohif/core/src';
 
-export default function PanelSegmentation({
-  servicesManager,
-  commandsManager,
-  children,
-}: withAppTypes) {
+export default function PanelSegmentation({ children }: withAppTypes) {
+  const { commandsManager, servicesManager } = useSystem();
   const { customizationService, displaySetService } = servicesManager.services;
 
   const { segmentationsWithRepresentations, disabled } =