+++ platform/app/src/components/ViewportGrid.tsx
@@ -258,6 +258,7 @@ function ViewerViewportGrid(props: withAppTypes) {
   }, [viewports, _getUpdatedViewports]);
 
   const onDropHandler = (viewportId, { displaySetInstanceUID }) => {
+    const { viewportGridService } = servicesManager.services;
     const customOnDropHandler = customizationService.getCustomization('customOnDropHandler');
     const dropHandlerPromise = customOnDropHandler({
       ...props,
@@ -265,7 +266,6 @@ function ViewerViewportGrid(props: withAppTypes) {
       displaySetInstanceUID,
       appConfig,
     });
-
     dropHandlerPromise.then(({ handled }) => {
       if (!handled) {
         const updatedViewports = _getUpdatedViewports(viewportId, displaySetInstanceUID);
@@ -273,6 +273,7 @@ function ViewerViewportGrid(props: withAppTypes) {
         commandsManager.run('setDisplaySetsForViewports', { viewportsToUpdate: updatedViewports });
       }
     });
+    viewportGridService.publishViewportOnDropHandled({ displaySetInstanceUID });
   };
 
   // Store previous isReferenceViewable values to avoid infinite loops